Move ascii_output function to taxcalc/utils.py file.
