+++ .coveragerc
@@ -2,7 +2,6 @@
 omit =
     taxcalc/tests/*
     conda.recipe/*
-    csv_to_ascii.py
     taxcalc/_version.py
     taxcalc/functions.py
     taxcalc/*.json
