+++ taxcalc/tests/test_utils.py
@@ -12,7 +12,6 @@ from pandas.util.testing import assert_series_equal
 CUR_PATH = os.path.abspath(os.path.dirname(__file__))
 sys.path.append(os.path.join(CUR_PATH, '..', '..'))
 from taxcalc import Policy, Records, Calculator
-from csv_to_ascii import ascii_output
 from taxcalc.utils import *
 
 # use 1991 PUF-like data to emulate current puf.csv, which is private
@@ -498,7 +497,7 @@ def asciifile():
     os.remove(f.name)
 
 
-def test_csv_to_ascii(csvfile, asciifile):
+def test_ascii_output_function(csvfile, asciifile):
     output_test = tempfile.NamedTemporaryFile(mode="a", delete=False)
     ascii_output(csv_results=csvfile.name, ascii_results=output_test.name)
     assert filecmp.cmp(output_test.name, asciifile.name)
